fix: переносит типы компонентов в отдельный файл
