+++ src/components/app/types.tsx
@@ -1,0 +1,3 @@
+export type AppProps = {
+  placesCount: number;
+}
