+++ src/components/blocks/place-cards/place-cards.tsx
@@ -1,19 +1,6 @@
+import { PlaceCardsProps } from './types';
+import {PlaceCardProps} from '../place-сard/types';
 import PlaceCard from '../place-сard/place-сard';
-
-export type PlaceCardProps = {
-  id: string;
-  title: string;
-  type: string;
-  price: number;
-  isFavorite?: boolean;
-  isPremium?: boolean;
-  rating: number;
-  previewImage: string;
-};
-
-export type PlaceCardsProps = {
-  cards: PlaceCardProps[];
-};
 
 function PlaceCards({cards}: PlaceCardsProps): JSX.Element {
   return (
