+++ src/components/blocks/place-cards/types.tsx
@@ -1,0 +1,5 @@
+import {PlaceCardProps} from '../place-сard/types';
+
+export type PlaceCardsProps = {
+  cards: PlaceCardProps[];
+}
