+++ src/components/blocks/place-сard/place-сard.tsx
@@ -1,6 +1,6 @@
 import {useState} from 'react';
 import {Link} from 'react-router-dom';
-import {PlaceCardProps} from '../place-cards/place-cards';
+import {PlaceCardProps} from './types';
 
 function PlaceCard(props: PlaceCardProps): JSX.Element {
   const [activeCardId, setActiveCardId] = useState('');
