+++ src/components/blocks/place-сard/types.tsx
@@ -1,0 +1,10 @@
+export type PlaceCardProps = {
+  id: string;
+  title: string;
+  type: string;
+  price: number;
+  isFavorite?: boolean;
+  isPremium?: boolean;
+  rating: number;
+  previewImage: string;
+}
