+++ src/components/blocks/private-route/private-route.tsx
@@ -1,10 +1,6 @@
+import {PrivateRouteProps} from './types';
 import {Navigate} from 'react-router-dom';
-import {AppRoute, AuthorizationStatus} from '../../const';
-
-type PrivateRouteProps = {
-  authorizationStatus: AuthorizationStatus;
-  children: JSX.Element;
-}
+import {AppRoute, AuthorizationStatus} from '../../consts';
 
 function PrivateRoute({authorizationStatus, children}: PrivateRouteProps): JSX.Element {
   return(
