+++ src/components/blocks/private-route/types.tsx
@@ -1,0 +1,6 @@
+import {AuthorizationStatus} from '../../consts';
+
+export type PrivateRouteProps = {
+  authorizationStatus: AuthorizationStatus;
+  children: JSX.Element;
+}
