+++ src/components/blocks/review/types.tsx
@@ -1,0 +1,11 @@
+export type ReviewProps = {
+  id: string;
+  date: string;
+  user: {
+    name: string;
+    avatarUrl: string;
+    isPro?: boolean;
+  };
+  comment: string;
+  rating: number;
+}
