+++ src/components/blocks/reviews/types.tsx
@@ -1,0 +1,5 @@
+import {ReviewProps} from '../review/types';
+
+export type ReviewsProps = {
+  items: ReviewProps[];
+}
